Refactor PersistentEngine and ScalarType to improve cohesion and reduce coupling.
- Move `PersistentEngine` to `relvar-storage/src/engine/mod.rs`.
- Derive `Ord` for `RelationType` and `TupleType`.
- Refactor `ScalarType::cmp` to delegate to `RelationType::cmp`.
- Update benchmarks to use correct imports.
