+++ relvar-storage/src/lib.rs
@@ -12,7 +12,7 @@
 
 #![warn(missing_docs)]
 
-pub mod persistent_engine;
+pub mod engine;
 pub mod storage;
 
 // WAL module is pub(crate) - not exposed to logical layer (TTM compliance)
@@ -21,7 +21,7 @@ pub(crate) mod wal;
 // MVCC module is pub(crate) - not exposed to logical layer (TTM compliance)
 pub(crate) mod mvcc;
 
-pub use persistent_engine::PersistentEngine;
+pub use engine::PersistentEngine;
 
 // Re-export key storage types
 pub use storage::{Catalog, CatalogError, HeapError, HeapFile, Page, PageError, PageFile};
